# perf: use multiple workers to read stdin