--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,7 +10,6 @@ assert2 = "0.3.14"
 clap = "4.5"
 compact_str = "0.7"
 criterion = "0.5"
-crossbeam = "0.8"
 either = "1.12"
 insta = "1.39"
 mimalloc = { version = "0.1", default-features = false }