Minor adjustments to nearly hit the 1.5 M/s milestone 🎉

30% faster than #86

```shell
$ cargo r --release --bin generator -- -n 5000000 -j 4 | cargo r --bin matching-engine --profile bench -- -j 4
    Finished `release` profile [optimized] target(s) in 0.08s
     Running `target/release/generator -n 5000000 -j 4`
    Finished `bench` profile [optimized + debuginfo] target(s) in 0.09s
     Running `target/release/matching-engine -j 4`
       Total 5 000 000 order(s) in 3.36s
     Average 1 487 083.35 orders/s

 Orderbook info
    Spread
     Ask 4531
     Bid 2805
    Length
     Ask 543698
     Bid 546012
```